backup/gynecology_chatbot_backup_20250519/backend/requirements.txt
# Django and REST framework
Django==5.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
drf-nested-routers==0.93.4

# Database
psycopg==3.1.16
psycopg-binary==3.1.16

# Image processing
Pillow==10.2.0

# API Clients
openai==1.3.7
google-generativeai==0.3.1
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Misc utilities
pydantic==2.6.1
